+++ sdk/securityinsight/arm-securityinsight/src/securityInsights.ts
@@ -38,6 +38,8 @@ import {
   ThreatIntelligenceIndicatorMetricsImpl,
   WatchlistsImpl,
   WatchlistItemsImpl,
+  ConfidentialWatchlistsImpl,
+  ConfidentialWatchlistItemsImpl,
   DataConnectorsImpl,
   DataConnectorsCheckRequirementsOperationsImpl,
   OperationsImpl
@@ -72,6 +74,8 @@ import {
   ThreatIntelligenceIndicatorMetrics,
   Watchlists,
   WatchlistItems,
+  ConfidentialWatchlists,
+  ConfidentialWatchlistItems,
   DataConnectors,
   DataConnectorsCheckRequirementsOperations,
   Operations
@@ -110,7 +114,7 @@ export class SecurityInsights extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-securityinsight/1.0.0-beta.2`;
+    const packageDetails = `azsdk-js-arm-securityinsight/1.0.0-beta.3`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -133,7 +137,7 @@ export class SecurityInsights extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2022-01-01-preview";
+    this.apiVersion = options.apiVersion || "2022-06-01-preview";
     this.alertRules = new AlertRulesImpl(this);
     this.actions = new ActionsImpl(this);
     this.alertRuleTemplates = new AlertRuleTemplatesImpl(this);
@@ -169,6 +173,8 @@ export class SecurityInsights extends coreClient.ServiceClient {
     );
     this.watchlists = new WatchlistsImpl(this);
     this.watchlistItems = new WatchlistItemsImpl(this);
+    this.confidentialWatchlists = new ConfidentialWatchlistsImpl(this);
+    this.confidentialWatchlistItems = new ConfidentialWatchlistItemsImpl(this);
     this.dataConnectors = new DataConnectorsImpl(this);
     this.dataConnectorsCheckRequirementsOperations = new DataConnectorsCheckRequirementsOperationsImpl(
       this
@@ -205,6 +211,8 @@ export class SecurityInsights extends coreClient.ServiceClient {
   threatIntelligenceIndicatorMetrics: ThreatIntelligenceIndicatorMetrics;
   watchlists: Watchlists;
   watchlistItems: WatchlistItems;
+  confidentialWatchlists: ConfidentialWatchlists;
+  confidentialWatchlistItems: ConfidentialWatchlistItems;
   dataConnectors: DataConnectors;
   dataConnectorsCheckRequirementsOperations: DataConnectorsCheckRequirementsOperations;
   operations: Operations;